requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
audioop-lts==0.2.2
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
docstring-parser==0.17.0
fastapi==0.121.3
ffmpy==1.0.0
filelock==3.20.0
fsspec==2025.10.0
gradio==5.50.0
gradio-client==1.14.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.1.5
idna==3.11
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
numpy==2.3.5
openai==2.8.1
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pydantic==2.12.3
pydantic-core==2.41.4
pydub==0.25.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
rich==14.2.0
ruff==0.14.6
safehttpx==0.1.7
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
